chore(sdk-coin-algo): move code from bitgo
- moves algo from bitgo to sdk-coin-algo

Ticket: BG-48143
