+++ modules/sdk-coin-algo/src/algoToken.ts
